+++ reproject/interpolation/__init__.py
@@ -2,4 +2,4 @@
 """
 Routines to carry out reprojection by interpolation.
 """
-from .core import *
+from .high_level import *
