+++ reproject/interpolation/core.py
@@ -9,76 +9,6 @@ from ..wcs_utils import convert_world_coordinates
 from ..array_utils import iterate_over_celestial_slices, pad_edge_1
 from ..utils import parse_input_data, parse_output_projection
 
-__all__ = ['reproject_interp']
-
-ORDER = {}
-ORDER['nearest-neighbor'] = 0
-ORDER['bilinear'] = 1
-ORDER['biquadratic'] = 2
-ORDER['bicubic']= 3
-
-
-def reproject_interp(input_data, output_projection, shape_out=None, hdu_in=None, order='bilinear'):
-    """
-    Reproject data to a new projection using interpolation (this is typically
-    the fastest way to reproject an image).
-
-    Parameters
-    ----------
-    input_data : str or `~astropy.io.fits.HDUList` or `~astropy.io.fits.PrimaryHDU` or `~astropy.io.fits.ImageHDU` or tuple
-        The input data to reproject. This can be:
-        
-            * The name of a FITS file
-            * An `~astropy.io.fits.HDUList` object
-            * An image HDU object such as a `~astropy.io.fits.PrimaryHDU` or
-              `~astropy.io.fits.ImageHDU`
-            * A tuple where the first element is a `~numpy.ndarray` and the
-              second element is either a `~astropy.wcs.WCS` or a
-              `~astropy.io.fits.Header` object
-
-    output_projection : `~astropy.wcs.WCS` or `~astropy.io.fits.Header`
-        The output projection, which can be either a `~astropy.wcs.WCS`
-        or a `~astropy.io.fits.Header` instance.
-    shape_out : tuple, optional
-        If ``output_projection`` is a `~astropy.wcs.WCS` instance, the
-        shape of the output data should be specified separately.
-    hdu_in : int or str, optional
-        If ``input_data`` is a FITS file or an ~astropy.io.fits.HDUList`
-        instance, specifies the HDU to use.
-    order : int or str, optional
-        The order of the interpolation (if ``mode`` is set to
-        ``'interpolation'``). This can be either one of the following strings:
-
-            * 'nearest-neighbor'
-            * 'bilinear'
-            * 'biquadratic'
-            * 'bicubic'
-
-        or an integer. A value of ``0`` indicates nearest neighbor
-        interpolation. 
-
-    Returns
-    -------
-    array_new : `~numpy.ndarray`
-        The reprojected array
-    footprint : `~numpy.ndarray`
-        Footprint of the input array in the output array. Values of 0 indicate
-        no coverage or valid values in the input image, while values of 1
-        indicate valid values.
-    """
-    
-    array_in, wcs_in = parse_input_data(input_data, hdu_in=hdu_in)
-    wcs_out, shape_out = parse_output_projection(output_projection, shape_out=shape_out)
-    
-    if isinstance(order, six.string_types):
-        order = ORDER[order]
-
-    # For now only celestial reprojection is supported
-    if wcs_in.has_celestial:
-        return _reproject_celestial(array_in, wcs_in, wcs_out, shape_out=shape_out, order=order)
-    else:
-        raise NotImplementedError("Currently only data with a WCS that includes a celestial component can be reprojected")
-
 
 def map_coordinates(image, coords, **kwargs):
